Add reprepro version obsolescence
We've moved reprepro to a version that supports keeping multiple
versions for the purpose of facilitating downgrades if people need it.
However the logic of the multiple versions functionality there does not
permit the granular rules we want. This implements them manually.

1. For Unstable builds, we only want to keep a max of 4 in the repo
(matching what we keep in the web archive). Since reprepro doesn't
support granularity of Limit at the component level, implement logic to
remove all but the most recent 4 unstables if there end up being more
than 4 after importing the new version.

2. For Stable builds, we want to keep the last point release of a given
major release (e.g. keep 10.10.13 but obsolete 10.10.12), while also
keeping older major versions (e.g. 10.9.10). Again the logic in reprepro
is too simplistic to handle this, so we implement it manually. For any
version that comes in, if the point version is above 0, we remove any
packages with a version that's one point release behind that one within
the same major version, while preserving everything else.